Feature : 게시물 조회 기능 완료

+++ wanted/src/main/java/com/wanted/post/controller/PostController.java
@@ -57,7 +57,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}")
-    private ResponseEntity<Object> findPost(@PathVariable("postId") Integer postId) {
+    private ResponseEntity<Object> findPost(@PathVariable("postId") Long postId) {
         log.info("findPost - Call");
 
         if (postId == null) {
